Second scenario fixed
In the first version of the solution, the clicking on the Continue/Confirm buttons didn't work from the second section. It was because even when not all the buttons are visible fr the user, all of them are in the DOM. So when I tried to click the second Continue button, it was actually clicking the first one.
I've changed the locators definition to include the sections which are introduced in the steps definition and passed as a parameters.